+++ common/src/main/java/dev/muon/medieval/hotbar/ManaProvider.java
@@ -1,8 +1,0 @@
-package dev.muon.medieval.hotbar;
-
-public interface ManaProvider {
-    double getCurrentMana();
-    float getMaxMana();
-    float getReservedMana(); // For use by Ars Nouveau, not sure if other mana systems have similar concepts
-    long getGameTime();  // For animation timing
-}