+++ common/src/main/java/dev/muon/medieval/hotbar/Position.java
@@ -1,7 +1,0 @@
-package dev.muon.medieval.hotbar;
-
-public record Position(int x, int y) {
-    public Position offset(int x, int y) {
-        return new Position(this.x + x, this.y + y);
-    }
-}